Improve Email::Valid validation
Email::Valid->address can return some modified string of original email
which may cause some confusion if treated as a boolean value.
One example is where it strips spaces and returns it if there is no other issues.
I've tried to handle this case, with an additional test case to make sure it fails.